+++ src/Integration/ModulesIntegration.php
@@ -17,7 +17,7 @@ use Sentry\State\Scope;
 final class ModulesIntegration implements IntegrationInterface
 {
     /**
-     * @var array The list of installed vendors
+     * @var array<string, string> The list of installed vendors
      */
     private static $loadedModules = [];
 
@@ -26,13 +26,13 @@ final class ModulesIntegration implements IntegrationInterface
      */
     public function setupOnce(): void
     {
-        Scope::addGlobalEventProcessor(function (Event $event) {
+        Scope::addGlobalEventProcessor(static function (Event $event): Event {
             $integration = SentrySdk::getCurrentHub()->getIntegration(self::class);
 
             // The integration could be bound to a client that is not the one
             // attached to the current hub. If this is the case, bail out
-            if ($integration instanceof self) {
-                self::applyToEvent($integration, $event);
+            if (null !== $integration) {
+                $integration->processEvent($event);
             }
 
             return $event;
@@ -44,8 +44,23 @@ final class ModulesIntegration implements IntegrationInterface
      *
      * @param self  $self  The instance of this integration
      * @param Event $event The event that will be enriched with the modules
+     *
+     * @deprecated since version 2.4, to be removed in 3.0
      */
     public static function applyToEvent(self $self, Event $event): void
+    {
+        @trigger_error(sprintf('The "%s" method is deprecated since version 2.4 and will be removed in 3.0.', __METHOD__), E_USER_DEPRECATED);
+
+        $self->processEvent($event);
+    }
+
+    /**
+     * Gathers information about the versions of the installed dependencies of
+     * the application and sets them on the event.
+     *
+     * @param Event $event The event
+     */
+    private function processEvent(Event $event): void
     {
         if (empty(self::$loadedModules)) {
             foreach (Versions::VERSIONS as $package => $rawVersion) {